chore: move Welcome sequence activity tests and remove WelcomeTestActivity
- Moved individual instrumented test files for `WelcomeOneActivity` through `WelcomeFiveActivity`.
- Removed `WelcomeTestActivity.kt` and its corresponding test file `WelcomeTestActivityTest.kt`.
