+++ app/src/androidTest/java/social/entourage/android/beforeLogin/welcome/WelcomeFiveActivityTest.kt
@@ -1,26 +1,25 @@
-package social.entourage.android.welcome
+package social.entourage.android.beforeLogin.welcome
 
-import androidx.test.ext.junit.rules.ActivityScenarioRule
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.intent.Intents
-import androidx.test.espresso.intent.Intents.intended
-import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.ext.junit.rules.ActivityScenarioRule
+import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import social.entourage.android.R
+import social.entourage.android.welcome.WelcomeFiveActivity
 
 @RunWith(AndroidJUnit4::class)
-class WelcomeTestActivityTest {
+class WelcomeFiveActivityTest {
 
     @get:Rule
-    val activityRule = ActivityScenarioRule(WelcomeTestActivity::class.java)
+    val activityRule = ActivityScenarioRule(WelcomeFiveActivity::class.java)
 
     @Before
     fun setUp() {
@@ -32,8 +31,9 @@ class WelcomeTestActivityTest {
         Intents.release()
     }
 
+
     @Test
-    fun test_activity_launches() {
-        // Activity is launched by the Rule
+    fun test_display() {
+        onView(withId(R.id.main_button)).check(matches(isDisplayed()))
     }
 }
