--- app/src/androidTest/java/social/entourage/android/welcome/WelcomeOneActivityTest.kt
@@ -1,45 +1,0 @@
-package social.entourage.android.welcome
-
-import androidx.test.ext.junit.rules.ActivityScenarioRule
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions.click
-import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.intent.Intents
-import androidx.test.espresso.intent.Intents.intended
-import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
-import androidx.test.espresso.matcher.ViewMatchers.withId
-import org.junit.After
-import org.junit.Before
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-
-@RunWith(AndroidJUnit4::class)
-class WelcomeOneActivityTest {
-
-    @get:Rule
-    val activityRule = ActivityScenarioRule(WelcomeOneActivity::class.java)
-
-    @Before
-    fun setUp() {
-        Intents.init()
-    }
-
-    @After
-    fun tearDown() {
-        Intents.release()
-    }
-
-
-    @Test
-    fun test_display() {
-        onView(withId(social.entourage.android.R.id.title_welcome_one)).check(matches(isDisplayed()))
-    }
-    @Test
-    fun test_close_button() {
-        onView(withId(social.entourage.android.R.id.close_button)).perform(click())
-        intended(hasComponent(social.entourage.android.MainActivity::class.java.name))
-    }
-}
